@@ -4,8 +4,8 @@ version: 0.0.3
 homepage: https://github.com/matheusperez/html_text
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
